feat(employee):_ids to Employee type
This commit adds a new field `customers_ids`` type in order to store the IDs of the customers assigned to. This change allows for of customer each. The `` the previous `` field, which stored an array of customer IDs. The `customers_ids` field will be used in the `EmployeeManagementPage to display and update the